+++ components/Player/Contexts/PresetProvider.tsx
@@ -1,17 +1,14 @@
 // PresetProvider.tsx
-import React, { useCallback, useContext, useEffect, useReducer, useState } from 'react';
+import React, { useContext, useReducer } from 'react';
 import { PlayerStateAction, playerStateReducer, PresetState } from './states';
-import { DEFAULT_VIDEO_ID, DEFAULT_VOLUME, GLOBAL_DISABLE_SAVE_PRESET } from '../../utils/DEFAULTS';
-import { clearPreset, loadPreset, savePreset } from './localStorageUtils';
+import { DEFAULT_VIDEO_ID, DEFAULT_VOLUME } from '../../utils/DEFAULTS';
+
 
 // ----------------- CONTEXT DECLARATION -----------------
 const PresetStateContext = React.createContext(
     {} as {
         presetState: PresetState;
         presetDispatch: React.Dispatch<PlayerStateAction>;
-        disablePersistPreset: boolean;
-        setDisablePersistPreset: React.Dispatch<React.SetStateAction<boolean>>;
-        clearPreset: () => void;
     },
 );
 
@@ -43,53 +40,12 @@ export function usePreset() {
 
 export default function PresetProvider({ children }: { children: React.ReactNode }) {
     const [presetState, presetDispatch] = useReducer(playerStateReducer, initialPresetState);
-    const [disablePersistPreset, setDisablePersistPreset] = useState(GLOBAL_DISABLE_SAVE_PRESET);
-
-    // ------- PRESET STATE PERSISTENCE -------
-    useEffect(() => {
-        presetDispatch({
-            type: 'setPreset',
-            payload: loadPreset(initialPresetState),
-        });
-    }, []);
-
-    const handleBeforeUnload = useCallback(
-        (e: BeforeUnloadEvent) => {
-            if (disablePersistPreset) return;
-            savePreset(presetState);
-            e.preventDefault();
-            e.returnValue = '';
-        },
-        [disablePersistPreset, presetState],
-    );
-
-    useEffect(() => {
-        window.addEventListener('beforeunload', handleBeforeUnload);
-
-        return () => {
-            window.removeEventListener('beforeunload', handleBeforeUnload);
-        };
-    }, [handleBeforeUnload]);
-
-    // ------- PAUSED TIMER -------
-
-    const setPausedAt = (pIndex: number, pPayload: number) => {
-        presetDispatch({
-            type: 'setPausedAt',
-            index: pIndex,
-            payload: pPayload,
-        });
-    };
-
 
     return (
         <PresetStateContext.Provider
             value={{
                 presetState,
                 presetDispatch,
-                disablePersistPreset,
-                setDisablePersistPreset,
-                clearPreset,
             }}
         >
             {children}
