+++ components/Player/Preferences/PreferencesSideMenu.tsx
@@ -1,0 +1,10 @@
+
+
+export function PreferencesSideMenu() {
+    return (
+        <>
+
+
+        </>
+    );
+}
